Used C++ style cast instead of C style cast  (#1628)

+++ moveit_experimental/kinematics_cache/v1/kinematics_cache/src/kinematics_cache.cpp
@@ -67,10 +67,10 @@ void KinematicsCache::setup(const KinematicsCache::Options& opt)
   cache_resolution_y_ = opt.resolution[1];
   cache_resolution_z_ = opt.resolution[2];
 
-  cache_size_x_ = (unsigned int)(opt.workspace_size[0] / opt.resolution[0]);
-  cache_size_y_ = (unsigned int)(opt.workspace_size[1] / opt.resolution[1]);
-  cache_size_z_ = (unsigned int)(opt.workspace_size[2] / opt.resolution[2]);
-  max_solutions_per_grid_location_ = std::max((unsigned int)1, opt.max_solutions_per_grid_location);
+  cache_size_x_ = static_cast<unsigned int>(opt.workspace_size[0] / opt.resolution[0]);
+  cache_size_y_ = static_cast<unsigned int>(opt.workspace_size[1] / opt.resolution[1]);
+  cache_size_z_ = static_cast<unsigned int>(opt.workspace_size[2] / opt.resolution[2]);
+  max_solutions_per_grid_location_ = std::max(1u, opt.max_solutions_per_grid_location);
   solution_dimension_ = joint_model_group_->getVariableCount();
   size_grid_node_ = max_solutions_per_grid_location_ * solution_dimension_;
   kinematics_cache_size_ = cache_size_x_ * cache_size_y_ * cache_size_z_;
